requirements.txt
anyio==4.9.0
APScheduler==3.11.0
asttokens==3.0.0
black==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
decorator==5.2.1
executing==2.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
ipython==9.0.2
ipython_pygments_lexers==1.1.1
isort==6.0.1
jedi==0.19.2
matplotlib-inline==0.1.7
mypy==1.15.0
mypy-extensions==1.0.0
packaging==24.2
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.3.7
pluggy==1.5.0
prompt_toolkit==3.0.50
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pytest==8.3.5
python-telegram-bot==22.0
requests==2.32.3
setuptools==77.0.3
sniffio==1.3.1
stack-data==0.6.3
traitlets==5.14.3
types-requests==2.32.0.20250328
typing_extensions==4.12.2
tzlocal==5.3.1
urllib3==2.3.0
wcwidth==0.2.13
